allow selecting environment to test
