+++ package.json
@@ -27,6 +27,6 @@
   },
   "dependencies": {
     "feather-icons": "^4.28.0",
-    "webnative": "^0.21.4"
+    "webnative": "^0.22.0-alpha-4"
   }
 }
